+++ my-app/src/App.jsx
@@ -1,10 +1,10 @@
-import TodoList from './components/TodoList';
+import {Todo} from './components/index';
 import './App.css';
 
 function App() {
   return (
     <div style={{margin: "2rem"}}>
-      <TodoList />
+      <Todo />
     </div>
   );
 }
