+++ my-app/src/components/index.js
@@ -1,0 +1,3 @@
+export {default as TodoList} from './TodoList'
+export {default as AddTodo} from './AddTodo'
+export {default as Todo} from './Todo'
